fbe-generator/requirements.txt
pandas
numpy
minio
Pillow